requirements.txt
fastmcp
python-dotenv
pandas
openai>=1.0.0 
azure-identity
azure-ai-agents
azure-ai-projects

pytest
pytest-asyncio
pytest-mock
